Split dev image

+++ hooks/post_gen_project.py
@@ -39,12 +39,12 @@ SYMLINKS_FILES = {
     "../../{{cookiecutter.deploy_project_dir}}/.ansible/playbooks/deploy_key_teardown.yml",  #noqa
     ".ansible/playbooks/site.yml":
     "../../{{cookiecutter.deploy_project_dir}}/.ansible/playbooks/site.yml",  #noqa
-    "tox.ini":    "{{cookiecutter.deploy_project_dir}}/tox.ini",  #noqa
+    # "tox.ini":    "{{cookiecutter.deploy_project_dir}}/tox.ini",  #noqa
     "Dockerfile": "{{cookiecutter.deploy_project_dir}}/Dockerfile",  #noqa
 }
 SYMLINKS = {}
 SYMLINKS.update(SYMLINKS_DIRS)
-SYMLINKS_DIRS.update(SYMLINKS_FILES)
+SYMLINKS.update(SYMLINKS_FILES)
 GITSCRIPT = """
 set -ex
 if [ ! -e .git ];then git init;fi
@@ -75,9 +75,11 @@ if [ ! -e "{{cookiecutter.deploy_project_dir}}/.git" ];then
 EGITSCRIPT = """
 {%raw%}vv() {{ echo "$@">&2;"$@"; }}{%endraw%}
 {% if cookiecutter.remove_cron %}
-rm -f crontab
-sed -i -re "/ADD .*cron/d" Dockerfile
-sed -i -re "/CMD .*cron/d" Dockerfile
+if [ -e Dockerfile ] && [ ! -h Dockerfile ];then
+    rm -f crontab
+    sed -i -re "/ADD .*cron/d" Dockerfile
+    sed -i -re "/CMD .*cron/d" Dockerfile
+fi
 {% endif %}
 {% for i in ['dev', 'prod', 'qa', 'staging'] -%}
 {% if not cookiecutter['{0}_host'.format(i)]%}
@@ -98,23 +100,31 @@ sed -i -re "/ADD private/d" Dockerfile
 sed -i -re "/ADD lib/d" Dockerfile
 rm -rf lib
 {% endif %}
+if [ -e Dockerfile ] && [ ! -h Dockerfile ];then
 sed -i -re \
 	"s/PY_VER=.*/PY_VER={{cookiecutter.py_ver}}/g" \
 	Dockerfile
 sed -i -re \
 	"s/project/{{cookiecutter.django_project_name}}/g" \
-	prod/*sh Dockerfile
+	Dockerfile
+fi
+if ( find prod/*sh 2>/dev/null );then
+sed -i -re \
+	"s/project/{{cookiecutter.django_project_name}}/g" \
+	prod/*sh
+fi
 set +x
+{% if not cookiecutter.use_submodule_for_deploy_code %}
 while read f;do
-echo $f
     if ( egrep -q "local/{{cookiecutter.app_type}}" "$f" );then
         echo "rewrite: $f"
-        vv sed -i -r \
+        vv echo sed -i -r \
         -e "s|{{cookiecutter.deploy_project_dir}}/||g" \
         -e "s|local/{{cookiecutter.app_type}}/||g" \
         "$f"
     fi
 done < <( find -type f|egrep -v "((^./(\.tox|\.git|local))|/static/)"; )
+{% endif %}
 set -x
         """
